extract AccountMenu from sidebar
